Changes for adding a model
The changes are for when adding a model it will prompt for the attributes if the appliance type requires/uses them.  Modifications to the Appliance class to cover the three different possibilities of input parameters depending on if the appliance has a repair plan, specifies capacity or max output, or has no attributes.  Other changes as required in other classes for supporting this functionality.